requirements.txt
# ============================================================
# ACD Framework — Production Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# ── Web framework ────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
python-multipart==0.0.9
httpx==0.27.0

# ── Pydantic ────────────────────────────────────────────────
pydantic==2.9.0
pydantic-settings==2.5.0

# ── Database ────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0          # SQLite async driver (dev)
asyncpg==0.30.0            # PostgreSQL async driver (prod)
alembic==1.13.3

# ── Cache / Queue ────────────────────────────────────────────
redis[asyncio]==5.1.0
celery[redis]==5.4.0
kombu==5.3.7
flower==2.0.1              # Celery monitoring UI

# ── Auth ────────────────────────────────────────────────────
PyJWT==2.9.0
bcrypt==4.2.0
slowapi==0.1.9             # Rate limiting

# ── RL / ML ─────────────────────────────────────────────────
stable-baselines3==2.3.2
gymnasium==0.29.1
torch>=2.2.0               # Install separately with CUDA build if needed
numpy==1.26.4
scipy==1.14.1

# ── Knowledge Graph ─────────────────────────────────────────
neo4j==5.25.0
sentence-transformers==3.1.1
faiss-cpu==1.9.0           # Use faiss-gpu for GPU support
scikit-learn==1.5.2

# ── LLM / Explainability ────────────────────────────────────
openai==1.51.0
anthropic==0.36.0          # Optional: Anthropic provider
tiktoken==0.8.0            # Token counting

# ── Config / Utilities ──────────────────────────────────────
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3
loguru==0.7.2

# ── Monitoring ──────────────────────────────────────────────
prometheus-client==0.21.0
sentry-sdk[fastapi,celery]==2.14.0

# ── HTTP client (for NVD API, ATT&CK download) ─────────────
urllib3==2.2.3
certifi==2024.8.30